+++ src/engine/search/mod.rs
@@ -5,9 +5,8 @@ use crate::engine::options::EngineOptions;
 use crate::engine::search::time_control::TimeStrategy;
 
 use crate::chess::game::Game;
-use crate::chess::player::Player;
-use crate::chess::square::Square;
 use crate::engine::search::move_provider::MoveProvider;
+use crate::engine::search::tables::HistoryTable;
 use crate::engine::search::transposition::SearchTranspositionTable;
 
 mod iterative_deepening;
@@ -15,6 +14,7 @@ mod move_ordering;
 pub mod move_provider;
 mod negamax;
 mod quiescence;
+mod tables;
 mod time_control;
 pub mod transposition;
 
@@ -37,14 +37,14 @@ mod params {
 
 pub struct PersistentState {
     pub tt: SearchTranspositionTable,
-    pub history: [[[i32; Square::N]; Square::N]; Player::N],
+    pub history_table: HistoryTable,
 }
 
 impl PersistentState {
     pub fn new() -> Self {
         Self {
             tt: SearchTranspositionTable::default(),
-            history: [[[0; Square::N]; Square::N]; Player::N],
+            history_table: HistoryTable::new(),
         }
     }
 }
@@ -199,15 +199,9 @@ pub fn search(
     time_strategy.init();
 
     persistent_state.tt.new_generation();
-
-    for from_square in 0..Square::N {
-        for to_square in 0..Square::N {
-            for player in 0..Player::N {
-                persistent_state.history[player][from_square][to_square] /=
-                    params::HISTORY_DECAY_FACTOR;
-            }
-        }
-    }
+    persistent_state
+        .history_table
+        .decay(params::HISTORY_DECAY_FACTOR);
 
     // The game is modified as moves are played during search. When the search terminates,
     // the game will be left in a dirty state since we will not undo the moves played to